@@ -2,6 +2,7 @@
   "name": "server",
   "version": "1.0.0",
   "description": "This project was bootstrapped with Fastify-CLI.",
+  "type": "module",
   "main": "app.js",
   "directories": {
     "test": "test"
@@ -10,32 +11,55 @@
     "test": "tap \"test/**/*.test.js\"",
     "start": "fastify start -l info app.js",
     "dev": "fastify start -w -l info -P app.js",
-    "pretest": "standard",
-    "lint": "eslint --fix",
-    "clean-db": "node ./scripts/clean-db.js"
+    "pretest": "npm run lint",
+    "lint": "eslint --fix ./**/*.js",
+    "clean-db": "node ./scripts/clean-db.js",
+    "seed-db": "node ./scripts/seed-db.js"
+  },
+  "husky": {
+    "hooks": {
+      "pre-commit": "npm run lint"
+    }
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "@fastify/autoload": "^5.0.0",
+    "@fastify/autoload": "^5.8.0",
+    "@fastify/cookie": "^9.1.0",
     "@fastify/cors": "^8.4.0",
+    "@fastify/csrf": "^6.2.0",
+    "@fastify/csrf-protection": "^6.3.0",
     "@fastify/env": "^4.2.0",
-    "@fastify/jwt": "^7.2.1",
-    "@fastify/oauth2": "^7.4.0",
-    "@fastify/sensible": "^5.0.0",
+    "@fastify/error": "^3.4.0",
+    "@fastify/helmet": "^11.1.1",
+    "@fastify/jwt": "^7.2.2",
+    "@fastify/oauth2": "^7.5.0",
+    "@fastify/sensible": "^5.5.0",
     "@fastify/session": "^10.5.0",
-    "@fastify/swagger": "^8.10.1",
-    "@thebguy/multer-gridfs-storage": "^5.0.2",
-    "fastify": "^4.23.2",
+    "@fastify/swagger": "^8.12.0",
+    "@fastify/swagger-ui": "^1.10.1",
+    "@thebguy/multer-gridfs-storage": "^5.0.3",
+    "connect-mongo": "^5.1.0",
+    "fastify": "^4.24.3",
     "fastify-cli": "^5.8.0",
     "fastify-multer": "^2.0.3",
-    "fastify-plugin": "^4.0.0",
-    "mongoose": "^7.5.3"
+    "fastify-plugin": "^4.5.1",
+    "mongoose": "^7.6.3"
   },
   "devDependencies": {
-    "eslint": "^8.51.0",
+    "@faker-js/faker": "^8.2.0",
+    "@istanbuljs/esm-loader-hook": "^0.2.0",
+    "eslint": "^8.52.0",
     "eslint-config-google": "^0.14.0",
-    "tap": "^16.1.0"
+    "husky": "^8.0.3",
+    "tap": "^18.5.2"
+  },
+  "tap": {
+    "node-arg": [
+      "--no-warnings",
+      "--experimental-loader",
+      "@istanbuljs/esm-loader-hook"
+    ]
   }
 }
