@@ -1,14 +1,13 @@
-'use strict';
+import fp from 'fastify-plugin';
+import multer from 'fastify-multer';
+import { GridFsStorage } from '@thebguy/multer-gridfs-storage';
 
-const fp = require('fastify-plugin');
-const multer = require('fastify-multer');
-const { GridFsStorage } = require('@thebguy/multer-gridfs-storage');
-
-module.exports = fp(async function (fastify, options) {
+export default fp(async function (fastify, options) {
   fastify.register(multer.contentParser);
 
   const storage = new GridFsStorage({
     url: fastify.config.MONGODB_URL,
+    // db: fastify.mongoose.connection.db,
     /**
      * @param {Request} req
      * @param {import("@thebguy/multer-gridfs-storage").GridFile} file
@@ -42,16 +41,6 @@ module.exports = fp(async function (fastify, options) {
     // Db is the database instance
     fastify.log.info('GRIDFS connection established!');
   });
-
-  // storage.on('streamError', (err) => {
-  //   // err is the error received from the GridFSBucketReadStream
-  //   fastify.log.error(err);
-  // });
-
-  // storage.on('dbError', (err) => {
-  //   // err is the error received from the MongoDB client
-  //   fastify.log.error(err);
-  // });
 
   // Decorate the Fastify instance
   fastify.decorate('upload', multer({ storage: storage }));
