@@ -1,12 +1,11 @@
-'use strict';
-const fp = require('fastify-plugin');
-
+import fp from 'fastify-plugin';
+import OAuth2 from '@fastify/oauth2';
 /**
  * registers the OAuth2 configuration for Google authentication.
  */
-module.exports = fp(async function (fastify, options) {
+export default fp(async function (fastify, options) {
   // Register the Fastify OAuth2 plugin.
-  fastify.register(require('@fastify/oauth2'), {
+  fastify.register(OAuth2, {
     name: 'googleOAuth2', // Name of the authentication strategy.
     scope: ['profile', 'email'], // The scope for the OAuth2 request. 'profile' gives access to basic profile information.
     credentials: {
@@ -14,9 +13,12 @@ module.exports = fp(async function (fastify, options) {
         id: process.env.GOOGLE_CLIENT_ID, // Client ID from environment variable.
         secret: process.env.GOOGLE_CLIENT_SECRET // Client Secret from environment variable.
       },
-      auth: require('@fastify/oauth2').GOOGLE_CONFIGURATION // Use the predefined Google OAuth2 configuration from Fastify OAuth2 plugin.
+      auth: (OAuth2).GOOGLE_CONFIGURATION // Use the predefined Google OAuth2 configuration from Fastify OAuth2 plugin.
     },
     startRedirectPath: '/auth/google', // Path to redirect user for authentication.
     callbackUri: 'http://localhost:3000/auth/google/callback' // The callback URL Google will redirect to after authentication.
   });
-}, { name: 'oauth' });
+}, {
+  name: 'oauth',
+  dependencies: ['session']
+});
