@@ -1,13 +1,12 @@
-'use strict';
+import https from 'https';
+import User from '../../models/User.js';
 
 /**
  * Route handlers for the Google OAuth2 callback.
  * @param {import("fastify").FastifyInstance} fastify
  * @param {Object} options plugin options, refer to https://www.fastify.io/docs/latest/Reference/Plugins/#plugin-options
  */
-module.exports = async function (fastify, options) {
-  const https = require('https');
-  const User = require('../../models/User');
+export default async function (fastify, options) {
   const getUserInfo = (accessToken) => {
     return new Promise((resolve, reject) => {
       const options = {
@@ -38,27 +37,91 @@ module.exports = async function (fastify, options) {
       req.end();
     });
   };
-  fastify.get('/google/callback', async function (request, reply) {
-    const token = await fastify.googleOAuth2.getAccessTokenFromAuthorizationCodeFlow(request);
-    const userInfo = await getUserInfo(token.token.access_token);
-    let user = await User.findOne({ gid: userInfo.id });
-    if (!user) {
-      user = new User({
-        gid: userInfo.id,
-        username: userInfo.name,
-        fullname: userInfo.name,
-        email: userInfo.email,
-        picture: userInfo.picture,
-      });
-      await user.save();
-    }
-    request.session.user = user;
-    reply.redirect('http://localhost:5173/');
+
+  fastify.get('/refresh', {
+    onRequest: fastify.csrfCheck,
+  }, async function (request, reply) {
+    await request.session.regenerate();
+    await reply.generateCsrf({ userInfo: request.session.user.fullname + request.session.user._id });
+
+    reply.setCookie('xsrf-t', request.session._csrf, {
+      secure: fastify.config.NODE_ENV === 'production',
+      httpOnly: true,
+      sameSite: true,
+      signed: true,
+      path: '/',
+    });
+    return { message: 'Refreshed' };
   });
 
-  fastify.post('/logout', async function (request, reply) {
-    request.session.destroy();
-    reply.send('Logged out');
+  fastify.post('/test', {
+    onRequest: fastify.csrfCheck,
+  }, async function (request, reply) {
+    return reply.send('Hello world');
   });
-  fastify.log.info('Oauth Route registered.');
-};
+
+  fastify.get('/google/callback', {
+    schema: {
+      tags: ['auth'],
+    },
+    async handler (request, reply) {
+      try {
+        const token = await fastify.googleOAuth2.getAccessTokenFromAuthorizationCodeFlow(request);
+        const userInfo = await getUserInfo(token.token.access_token);
+        let user = await User.findOne({ gid: userInfo.id });
+
+        if (!user) {
+          user = new User({
+            gid: userInfo.id,
+            username: userInfo.name,
+            fullname: userInfo.name,
+            email: userInfo.email,
+            profilePhoto: userInfo.picture,
+          });
+          await user.save();
+        }
+
+        request.session.user = user;
+        await reply.generateCsrf({ userInfo: user.fullname.replace(/\s/g, '') + user._id.toString() });
+
+        reply.setCookie('xsrf-t', request.session._csrf, {
+          secure: fastify.config.NODE_ENV === 'production',
+          httpOnly: true,
+          sameSite: true,
+          signed: true,
+          path: '/',
+        });
+
+        return reply.redirect('http://localhost:5173/');
+      } catch (err) {
+        fastify.log.error(err);
+        throw new Error('Internal Server Error');
+      }
+    },
+  });
+
+  fastify.post('/logout', {
+    schema: {
+      tags: ['auth'],
+      response: {
+        200: {
+          type: 'string',
+          description: 'Logout message',
+        },
+      },
+    },
+    async handler (request, reply) {
+      try {
+        await request.session.destroy();
+        reply.clearCookie('sid');
+        reply.clearCookie('xsrf-t');
+        reply.clearCookie('oauth2-redirect-state');
+
+        return reply.send('Logged out');
+      } catch (err) {
+        fastify.log.error(err);
+        throw new Error('Internal Server Error');
+      }
+    },
+  });
+}
